+++ tests/test_example.py
@@ -13,17 +13,18 @@ class TestExample(TestCase):
     @chdir_context("example")
     def test_run(self) -> None:
         """运行所有示例脚本"""
-        for filepath in Path(".").glob("*.py"):
-            if filepath.stem in {"clabel", "river", "nerv_style"}:
+        for file_path in Path(".").glob("*.py"):
+            file_stem = file_path.stem
+            if file_stem in {"clabel", "river", "nerv_style"}:
                 continue
             try:
                 subprocess.run(
-                    args=["python", str(filepath)],
+                    args=["python", str(file_path)],
                     check=True,
                     capture_output=True,
                     text=True,
                 )
-                print(f"[OK] {filepath.stem}")
+                print(f"[OK] {file_stem}")
             except CalledProcessError as e:
-                print(f"[FAIL] {filepath.stem}")
+                print(f"[FAIL] {file_stem}")
                 print(e.stderr)